with $app in index.php
