+++ hw4/composer.json
@@ -1,0 +1,16 @@
+{
+    "name": "elena/php_2023",
+    "type": "project",
+    "autoload": {
+        "psr-4": {
+            "Elena\\Php2023\\": "src/"
+        }
+    },
+    "authors": [
+        {
+            "name": "Елена Спиктаренко",
+            "email": "elena.spiktarenko@gmail.com"
+        }
+    ],
+    "require": {}
+}
