+++ hw4/docker_/fpm/conf.d/app.conf
@@ -1,0 +1,10 @@
+upstream memcached_backend {
+    server memcached-1:11211;
+    server memcached-2:11211;
+}
+
+server {
+    location /fastgi/ {
+    fastcgi_pass memcached_backend;
+   }
+}
