+++ hw4/docker_/nginx-balancer/conf.d/nginx.conf
@@ -10,7 +10,6 @@ index index.php index.html;
 server {
          listen 80;
          server_name mysite.local;
-
          location / {
             proxy_pass   http://nginx-balancer;
             proxy_set_header Host $host;
