+++ hw4/docker_/nginx/conf.d/nginx.conf
@@ -6,8 +6,8 @@
 
 server {
     listen 80;
-    root /var/www/public;
-    index index.php ;
+    root /var/www;
+    index /public/index.php ;
 
     location ~* .(jpg|jpeg|gif|css|png|js|ico|html)$ {
             access_log off;
@@ -15,14 +15,14 @@ server {
         }
 
     location / {
-        try_files $uri $uri/ /index.php?$query_string;
+        try_files $uri $uri/ /public/index.php?$query_string;
         }
 
     location ~ \.php$ {
     try_files $uri =404;
     fastcgi_split_path_info ^(.+\.php)(/.+)$;
     fastcgi_pass fpm-php;
-    fastcgi_index index.php;
+    fastcgi_index /public/index.php;
     include fastcgi_params;
     fastcgi_param SCRIPT_FILENAME $document_root$fastcgi_script_name;
     fastcgi_param PATH_INFO $fastcgi_path_info;
