+++ hw5/public/index.html
@@ -1,0 +1,4 @@
+<?php
+echo "Hello, Otus!";
+
+phpinfo();
